requirements.txt
torch
pillow
open-clip-torch
transformers
pyyaml
sentencepiece
clip
pandas